Handle dragging with NVDA/JAWS in browse mode
