+++ packages/@react-aria/interactions/src/useFocusVisible.ts
@@ -15,8 +15,7 @@
 // NOTICE file in the root directory of this source tree.
 // See https://github.com/facebook/react/tree/cc7c1aece46a6b69b41958d731e0fd27c94bfc6c/packages/react-interactions
 
-import {isMac} from '@react-aria/utils';
-import {isVirtualClick} from './utils';
+import {isMac, isVirtualClick} from '@react-aria/utils';
 import {useEffect, useState} from 'react';
 
 export type Modality = 'keyboard' | 'pointer' | 'virtual';
