+++ packages/@react-aria/interactions/src/utils.ts
@@ -13,26 +13,6 @@
 import {FocusEvent as ReactFocusEvent, useCallback, useRef} from 'react';
 import {useLayoutEffect} from '@react-aria/utils';
 
-// Original licensing for the following method can be found in the
-// NOTICE file in the root directory of this source tree.
-// See https://github.com/facebook/react/blob/3c713d513195a53788b3f8bb4b70279d68b15bcc/packages/react-interactions/events/src/dom/shared/index.js#L74-L87
-
-// Keyboards, Assistive Technologies, and element.click() all produce a "virtual"
-// click event. This is a method of inferring such clicks. Every browser except
-// IE 11 only sets a zero value of "detail" for click events that are "virtual".
-// However, IE 11 uses a zero value for all click events. For IE 11 we rely on
-// the quirk that it produces click events that are of type PointerEvent, and
-// where only the "virtual" click lacks a pointerType field.
-
-export function isVirtualClick(event: MouseEvent | PointerEvent): boolean {
-  // JAWS/NVDA with Firefox.
-  if ((event as any).mozInputSource === 0 && event.isTrusted) {
-    return true;
-  }
-
-  return event.detail === 0 && !(event as PointerEvent).pointerType;
-}
-
 export class SyntheticFocusEvent implements ReactFocusEvent {
   nativeEvent: FocusEvent;
   target: Element;
